# PERF: GroupBy.any/all operate blockwise instead of column-wise