--- a/pandas/_libs/groupby.pyx
+++ b/pandas/_libs/groupby.pyx
@@ -443,24 +443,17 @@ def group_any_all(int8_t[:, ::1] out,
     else:
         raise ValueError("'bool_func' must be either 'any' or 'all'!")
 
-    if mask is not None and mask.ndim == 2 and (<object>mask).shape[1] != 1:
-        # FIXME: in some builds mask.shape is e.g. [1, 2, 0, 0, 0, 0, 0, 0]
-        #  seems like a problem in cython
-        #  github.com/cython/cython/issues/4321
-        raise NotImplementedError(
-            "group_any_all assumes mask.shape[1] == 1", mask.shape
-        )
-
     out[:] = 1 - flag_val
 
     with nogil:
         for i in range(N):
             lab = labels[i]
-            if lab < 0 or (skipna and mask[i, 0]):
-                # NB: mask[i, 0] here is assuming mask.shape[1] == 1
+            if lab < 0:
                 continue
 
             for j in range(K):
+                if skipna and mask[i, j]:
+                    continue
 
                 if nullable and mask[i, j]:
                     # Set the position as masked if `out[lab] != flag_val`, which