--- a/pandas/core/groupby/generic.py
+++ b/pandas/core/groupby/generic.py
@@ -1614,12 +1614,15 @@ class DataFrameGroupBy(GroupBy[DataFrame]):
         -------
         DataFrame
         """
-        indexed_output = {key.position: val for key, val in output.items()}
-        columns = Index([key.label for key in output])
-        columns._set_names(self._obj_with_exclusions._get_axis(1 - self.axis).names)
+        if isinstance(output, DataFrame):
+            result = output
+        else:
+            indexed_output = {key.position: val for key, val in output.items()}
+            columns = Index([key.label for key in output])
+            columns._set_names(self._obj_with_exclusions._get_axis(1 - self.axis).names)
 
-        result = self.obj._constructor(indexed_output)
-        result.columns = columns
+            result = self.obj._constructor(indexed_output)
+            result.columns = columns
 
         if not self.as_index:
             self._insert_inaxis_grouper_inplace(result)