--- a/pandas/core/groupby/groupby.py
+++ b/pandas/core/groupby/groupby.py
@@ -1520,13 +1520,13 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
         return self.obj._constructor
 
     @final
-    def _bool_agg(self, val_test, skipna):
+    def _bool_agg(self, val_test: Literal["any", "all"], skipna: bool):
         """
         Shared func to call any / all Cython GroupBy implementations.
         """
 
         def objs_to_bool(vals: ArrayLike) -> tuple[np.ndarray, type]:
-            if is_object_dtype(vals):
+            if is_object_dtype(vals.dtype):
                 # GH#37501: don't raise on pd.NA when skipna=True
                 if skipna:
                     vals = np.array([bool(x) if not isna(x) else True for x in vals])
@@ -1535,7 +1535,7 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
             elif isinstance(vals, BaseMaskedArray):
                 vals = vals._data.astype(bool, copy=False)
             else:
-                vals = vals.astype(bool)
+                vals = vals.astype(bool, copy=False)
 
             return vals.view(np.int8), bool
 
@@ -1555,6 +1555,7 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
             numeric_only=False,
             cython_dtype=np.dtype(np.int8),
             needs_values=True,
+            needs_2d=True,
             needs_mask=True,
             needs_nullable=True,
             pre_processing=objs_to_bool,
@@ -2910,16 +2911,24 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
         if min_count is not None:
             base_func = partial(base_func, min_count=min_count)
 
+        real_2d = how in ["group_any_all"]
+
         def blk_func(values: ArrayLike) -> ArrayLike:
+            values = values.T
+            ncols = 1 if values.ndim == 1 else values.shape[1]
+
             if aggregate:
                 result_sz = ngroups
             else:
-                result_sz = len(values)
+                result_sz = values.shape[-1]
 
             result: ArrayLike
-            result = np.zeros(result_sz, dtype=cython_dtype)
+            result = np.zeros(result_sz * ncols, dtype=cython_dtype)
             if needs_2d:
-                result = result.reshape((-1, 1))
+                if real_2d:
+                    result = result.reshape((result_sz, ncols))
+                else:
+                    result = result.reshape(-1, 1)
             func = partial(base_func, out=result)
 
             inferences = None
@@ -2934,12 +2943,14 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
                     vals, inferences = pre_processing(vals)
 
                 vals = vals.astype(cython_dtype, copy=False)
-                if needs_2d:
+                if needs_2d and vals.ndim == 1:
                     vals = vals.reshape((-1, 1))
                 func = partial(func, values=vals)
 
             if needs_mask:
                 mask = isna(values).view(np.uint8)
+                if needs_2d and mask.ndim == 1:
+                    mask = mask.reshape(-1, 1)
                 func = partial(func, mask=mask)
 
             if needs_nullable:
@@ -2948,11 +2959,14 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
 
             func(**kwargs)  # Call func to modify indexer values in place
 
-            if needs_2d:
-                result = result.reshape(-1)
-
             if result_is_index:
                 result = algorithms.take_nd(values, result, fill_value=fill_value)
+
+            if real_2d and values.ndim == 1:
+                assert result.shape[1] == 1, result.shape
+                result = result[:, 0]
+                if needs_mask:
+                    mask = mask[:, 0]
 
             if post_processing:
                 pp_kwargs = {}
@@ -2961,7 +2975,26 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
 
                 result = post_processing(result, inferences, **pp_kwargs)
 
-            return result
+            if needs_2d and not real_2d:
+                if result.ndim == 2:
+                    assert result.shape[1] == 1
+                    result = result[:, 0]
+
+            return result.T
+
+        obj = self._obj_with_exclusions
+        if obj.ndim == 2 and self.axis == 0 and needs_2d and real_2d:
+
+            mgr = obj._mgr
+            if numeric_only:
+                mgr = mgr.get_numeric_data()
+
+            res_mgr = mgr.grouped_reduce(blk_func, ignore_failures=True)
+            output = type(obj)(res_mgr)
+            if aggregate:
+                return self._wrap_aggregated_output(output)
+            else:
+                return self._wrap_transformed_output(output)
 
         error_msg = ""
         for idx, obj in enumerate(self._iterate_slices()):