--- a/pandas/core/groupby/groupby.py
+++ b/pandas/core/groupby/groupby.py
@@ -2997,7 +2997,8 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
             if numeric_only:
                 mgr = mgr.get_numeric_data()
 
-            res_mgr = mgr.grouped_reduce(blk_func, ignore_failures=True)
+            # setting ignore_failures=False for troubleshooting
+            res_mgr = mgr.grouped_reduce(blk_func, ignore_failures=False)
             output = type(obj)(res_mgr)
             if aggregate:
                 return self._wrap_aggregated_output(output)