@jbrockmendel 2021-08-03
> Why are we moving to 2d kernels here at all?

For perf, see benchmarks posted in OP.

> I would rather NOT have 2d and 1d kernels for things.

my thought is that everything going through _get_cython_result here should be blockwise, which will both improve perf and let us de-duplicate with the existing groupby-blockwise code