@jbrockmendel 2021-08-04
From #38992:

> i suppose it makes sense to support both 1d and 2d kernels on things. it does lead to some code duplication, but performance can be great if we don't need to operate column-by-column all the time. However the codebase is mostly 1d currently, with some efforts to add 2d kernels.
>
> does this concur with your thinking?

1 ) I don't think its accurate to say the [kernels are] mostly 1d currently.  Looking at groupby.pyx python-visible functions i see 15 that take 2D inputs and 4 that take 1D inputs. (in master, not this PR)

2) Some of the perf benefits to operating blockwise vs columnwise are in surprising places.  e.g. in this branch I found that the pre_processor call to `.astype(bool)` is a major time-sink when done column-wise.

3) some of this solves itself if we ever move to numba.