@jbrockmendel 2021-08-17
gentle ping, nearing completion on a follow-up that does the same perf-improvement for std